Clean up Name Requests API implementation. Split name_requests controller into two files, one for collection ops and one for entity ops. Implement a rudimentary PATCH in the NameRequest resource. Clean up Swagger API docs.
