+++ api/namex/resources/name_requests/__init__.py
@@ -1,1 +1,2 @@
-from .name_requests import api, NameRequest, NameRequests
+from .name_request import NameRequest
+from .name_requests import NameRequests
