+++ api/namex/resources/name_requests/api_namespace.py
@@ -1,0 +1,4 @@
+from flask_restplus import Namespace
+
+# Register a local namespace for the NR reserve
+api = Namespace('nameRequest', description='API for Public Facing Name Requests')
