+++ api/namex/resources/name_requests/configuration.py
@@ -1,0 +1,6 @@
+import os
+
+SOLR_CORE = 'possible.conflicts'
+
+SOLR_URL = os.getenv('SOLR_BASE_URL')
+SOLR_API_URL = SOLR_URL + '/solr/'
